Move water.py to fluid.py since it also contains UnitFluid
